requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Image processing
pillow==11.0.0

# HTTP client
httpx==0.27.0

# Document generation (Phase 2 - PrimitivesPipe)
# python-docx==1.1.0      # DOCX generation
# python-pptx==1.0.0      # PPTX generation
# weasyprint==62.0        # PDF generation (requires system deps)
# mistune==3.0.2          # Markdown parsing
